+++ actix-http/src/service.rs
@@ -1,5 +1,4 @@
 use std::{
-    error::Error as StdError,
     fmt,
     future::Future,
     marker::PhantomData,
@@ -58,7 +57,6 @@ where
     S::Response: Into<Response<B>> + 'static,
     <S::Service as Service<Request>>::Future: 'static,
     B: MessageBody + 'static,
-    B::Error: Into<Box<dyn StdError>>,
 {
     /// Create new `HttpService` instance.
     pub fn new<F: IntoServiceFactory<S, Request>>(service: F) -> Self {
@@ -157,7 +155,6 @@ where
     <S::Service as Service<Request>>::Future: 'static,
 
     B: MessageBody + 'static,
-    B::Error: Into<Box<dyn StdError>>,
 
     X: ServiceFactory<Request, Config = (), Response = Request>,
     X::Future: 'static,
@@ -214,7 +211,6 @@ mod openssl {
         <S::Service as Service<Request>>::Future: 'static,
 
         B: MessageBody + 'static,
-        B::Error: Into<Box<dyn StdError>>,
 
         X: ServiceFactory<Request, Config = (), Response = Request>,
         X::Future: 'static,
@@ -287,7 +283,6 @@ mod rustls {
         <S::Service as Service<Request>>::Future: 'static,
 
         B: MessageBody + 'static,
-        B::Error: Into<Box<dyn StdError>>,
 
         X: ServiceFactory<Request, Config = (), Response = Request>,
         X::Future: 'static,
@@ -354,7 +349,6 @@ where
     <S::Service as Service<Request>>::Future: 'static,
 
     B: MessageBody + 'static,
-    B::Error: Into<Box<dyn StdError>>,
 
     X: ServiceFactory<Request, Config = (), Response = Request>,
     X::Future: 'static,
@@ -491,7 +485,6 @@ where
     S::Response: Into<Response<B>> + 'static,
 
     B: MessageBody + 'static,
-    B::Error: Into<Box<dyn StdError>>,
 
     X: Service<Request, Response = Request>,
     X::Error: Into<Response<BoxBody>>,
@@ -560,7 +553,6 @@ pin_project! {
         S::Error: Into<Response<BoxBody>>,
 
         B: MessageBody,
-        B::Error: Into<Box<dyn StdError>>,
 
         X: Service<Request, Response = Request>,
         X::Error: Into<Response<BoxBody>>,
@@ -597,7 +589,6 @@ pin_project! {
         S::Response: 'static,
 
         B: MessageBody,
-        B::Error: Into<Box<dyn StdError>>,
 
         X: Service<Request, Response = Request>,
         X::Error: Into<Response<BoxBody>>,
@@ -620,7 +611,6 @@ where
     S::Response: Into<Response<B>> + 'static,
 
     B: MessageBody + 'static,
-    B::Error: Into<Box<dyn StdError>>,
 
     X: Service<Request, Response = Request>,
     X::Error: Into<Response<BoxBody>>,