+++ src/response/response.rs
@@ -1,6 +1,5 @@
 use std::{
     cell::{Ref, RefMut},
-    error::Error as StdError,
     fmt,
     future::Future,
     mem,
@@ -241,7 +240,6 @@ impl<B> HttpResponse<B> {
     pub fn map_into_boxed_body(self) -> HttpResponse<BoxBody>
     where
         B: MessageBody + 'static,
-        B::Error: Into<Box<dyn StdError + 'static>>,
     {
         self.map_body(|_, body| BoxBody::new(body))
     }