+++ src/service.rs
@@ -1,6 +1,5 @@
 use std::{
     cell::{Ref, RefMut},
-    error::Error as StdError,
     fmt, net,
     rc::Rc,
 };
@@ -426,7 +425,6 @@ impl<B> ServiceResponse<B> {
     pub fn map_into_boxed_body(self) -> ServiceResponse<BoxBody>
     where
         B: MessageBody + 'static,
-        B::Error: Into<Box<dyn StdError + 'static>>,
     {
         self.map_body(|_, body| BoxBody::new(body))
     }