+++ src/types/either.rs
@@ -14,7 +14,7 @@ use pin_project_lite::pin_project;
 use crate::{
     body, dev,
     web::{Form, Json},
-    BoxError, Error, FromRequest, HttpRequest, HttpResponse, Responder,
+    Error, FromRequest, HttpRequest, HttpResponse, Responder,
 };
 
 /// Combines two extractor or responder types into a single type.
@@ -144,9 +144,7 @@ impl<L, R> Either<L, R> {
 impl<L, R> Responder for Either<L, R>
 where
     L: Responder,
-    <L::Body as dev::MessageBody>::Error: Into<BoxError>,
     R: Responder,
-    <R::Body as dev::MessageBody>::Error: Into<BoxError>,
 {
     type Body = body::EitherBody<L::Body, R::Body>;