+++ CHANGES.md
@@ -4,7 +4,7 @@
 ### Added
 * Methods on `AcceptLanguage`: `ranked` and `preference`. [#2480]
 * `AcceptEncoding` typed header. [#2482]
-* `HttpResponse::map_into_boxed_body`. [#????]
+* `HttpResponse::map_into_boxed_body`. [#2468]
 
 ### Changed
 * Rename `Accept::{mime_precedence => ranked}`. [#2480]
@@ -14,7 +14,7 @@
 * Accept wildcard `*` items in `AcceptLanguage`. [#2480]
 * Typed headers containing lists that require one or more items now enforce this minimum. [#2482]
 
-[#????]: https://github.com/actix/actix-web/pull/????
+[#2468]: https://github.com/actix/actix-web/pull/2468
 [#2480]: https://github.com/actix/actix-web/pull/2480
 [#2482]: https://github.com/actix/actix-web/pull/2482